% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{Get useful info on all countries}
\usage{
countries()
}
\value{
A dataframe.
}
\description{
Get a dataframe of useful information on all countries such as
FIFA codes, Dial codes, ISO codes, currency, official name, and a whole lot
more. See https://datahub.io/core/country-codes
}
