% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddg.R
\name{ddg}
\alias{ddg}
\title{Duckduckgo Instant-search}
\usage{
ddg(x, max_rel)
}
\arguments{
\item{x}{The search string}

\item{max_rel}{(optional) Max. number of related results (default 3), or the
number of results returned, which ever is smaller}
}
\value{
The search URL used
}
\description{
Perform a DDG Instant search on a query term.
}
\examples{
ddg('ngrams')
ddg('ngrams', 2)
}
