% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\name{dfetch}
\alias{dfetch}
\title{Fetch/load a dataset - see dsearch() and dinfo()}
\usage{
dfetch(pkg, ds)
}
\arguments{
\item{pkg}{The package name}

\item{ds}{The dataset name}
}
\value{
A vector of all the loaded dataset objects. Console output will list
all found tables in the dataset (usually just the one)
}
\description{
Get and load a dataset in a package. Use dsearch('query') to
find packages with datasets, and dinfo('package') to get the package info
and datasets.
}
\examples{
dfetch('TSrepr', 'elec_load.rda')
}
