% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\name{dsearch}
\alias{dsearch}
\title{Search for a dataset in all CRAN packages}
\usage{
dsearch(val)
}
\arguments{
\item{val}{The search string}
}
\value{
No return, outputs a list of packages to console.
Note: some datasets will have package dependencies to load, please take heed
of any such messages when viewing the loaded dataset object.
}
\description{
Search for, and fetch datasets in -all- CRAN packages, installed
or not.
Note: file types may vary, use the full file name of the dataset you require,
as displayed. Package & dataset names are case-sensitive.
}
\examples{
dsearch('Time series')
}
