% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\docType{data}
\name{elec_load}
\alias{elec_load}
\title{A global var.}
\format{
An object of class \code{character} of length 1.
}
\usage{
elec_load
}
\description{
A global var for the package
}
\keyword{internal}
