% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\docType{data}
\name{sexout}
\alias{sexout}
\title{A global var.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
sexout
}
\description{
A global var for the package
}
\keyword{internal}
