% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharebin.R
\name{sharebin}
\alias{sharebin}
\title{"Paste" a file for web sharing}
\usage{
sharebin(file_name)
}
\arguments{
\item{file_name}{(Optional) - `sharebin()` will paste the file in the current
active editor window. Otherwise `sharebin('file.suff')`, where 'file.suff'
is a file name that must exist in the working dir (getwd()) and is TEXT based
 (.R, .Rmd, .csv, or any other code or text file)
 ~~(DO NOT POST SENSITIVE DATA!)~~}
}
\value{
A URL you can share.
}
\description{
"Paste" a TEXT based file to a "bin" on the web for sharing.
~~(DO NOT POST SENSITIVE DATA!)~~.
}
