% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\docType{data}
\name{webtools}
\alias{webtools}
\title{A global var.}
\format{
An object of class \code{environment} of length 0.
}
\usage{
webtools
}
\description{
A global var for the package
}
\keyword{internal}
