% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yada.R
\name{yadayada}
\alias{yadayada}
\title{Get two random Seinfeld quotes in a row (guessing game)}
\usage{
yadayada()
}
\value{
A vector of the two quotes
}
\description{
Get a random Seinfeld quote, followed by the very next line.
Designed to be played in a terminal/console, otherwise it just returns the
consecutive quotes.
}
\examples{
yadayada()
}
