% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddg.R
\name{ddg}
\alias{ddg}
\title{Duckduckgo Instant-search}
\usage{
ddg(q, max_rel, get_result)
}
\arguments{
\item{q}{The search string}

\item{max_rel}{(optional) Max. number of related results (default 3), or the
number of results returned, which ever is smaller}

\item{get_result}{(optional) Return the first result (either the `Instant
Result`` or, if none, the first related result)}
}
\value{
A "safe" search URL, or the first result if get_result=TRUE passed
}
\description{
Perform a DDG Instant search on a query term.
}
\examples{
ddg('ngrams')
ddg('ngrams', 2)
ddg('ngrams', get_result=TRUE)
}
