% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R
\name{dsearch}
\alias{dsearch}
\title{Search for a dataset in all CRAN packages}
\usage{
dsearch(query)
}
\arguments{
\item{query}{The search string}
}
\value{
A tibble of packages (having datasets) matching the search query.
}
\description{
Search for, and fetch datasets in -all- CRAN packages, installed
or not.
Note: file types may vary, use the full file name of the dataset you require,
as displayed. Package & dataset names are case-sensitive.
}
\examples{
dsearch('Time series')
}
