% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsearch.R, R/yada.R
\docType{data}
\name{webtools}
\alias{webtools}
\title{A global var.}
\format{
An object of class \code{environment} of length 1.

An object of class \code{environment} of length 1.
}
\usage{
webtools

webtools
}
\description{
A global var for the package

A global var for the package
}
\keyword{internal}
